add artificial items electirizer and magmarizer
