+++ app/models/colyseus-models/count.ts
@@ -28,4 +28,5 @@ export default class Count extends Schema implements ICount {
   @type("uint8") soulDewCount = 0
   @type("uint8") healOrderCount = 0
   @type("uint8") attackOrderCount = 0
+  @type("uint8") magmarizerCount = 0
 }
