+++ app/public/src/game/components/item-detail.css
@@ -56,7 +56,7 @@
   z-index: 2;
 }
 
-.game-item-detail-description:last-child {
+.react-tooltip .game-item-detail-description:last-child {
   margin-bottom: -5px;
 }
