+++ app/public/src/game/components/item-detail.tsx
@@ -2,8 +2,8 @@ import { GameObjects } from "phaser"
 import React, { useMemo } from "react"
 import ReactDOM from "react-dom/client"
 import { Tooltip } from "react-tooltip"
-import { ItemRecipe, ItemStats } from "../../../../types/Config"
-import { Item } from "../../../../types/enum/Item"
+import { ItemStats } from "../../../../types/Config"
+import { Item, ItemRecipe } from "../../../../types/enum/Item"
 import { addIconsToDescription } from "../../pages/utils/descriptions"
 import "./item-detail.css"
 import { useTranslation } from "react-i18next"
