+++ app/public/src/game/scenes/game-scene.ts
@@ -6,7 +6,7 @@ import UnownManager from "../components/unown-manager"
 import WeatherManager from "../components/weather-manager"
 import ItemsContainer from "../components/items-container"
 import Pokemon from "../components/pokemon"
-import { Dungeon, DungeonPMDO, ItemRecipe } from "../../../../types/Config"
+import { Dungeon, DungeonPMDO } from "../../../../types/Config"
 import firebase from "firebase/compat/app"
 import { transformCoordinate } from "../../pages/utils/utils"
 import { Room } from "colyseus.js"
@@ -22,7 +22,7 @@ import {
   Transfer
 } from "../../../../types"
 import { DesignTiled } from "../../../../core/design"
-import { Item } from "../../../../types/enum/Item"
+import { Item, ItemRecipe } from "../../../../types/enum/Item"
 import Player from "../../../../models/colyseus-models/player"
 import MinigameManager from "../components/minigame-manager"
 import LoadingManager from "../components/loading-manager"
